+++ src/pipeline/Middleware.js
@@ -1,0 +1,9 @@
+export default class Middleware {
+  constructor (next) {
+    this.next = next
+  }
+
+  pipe (request) {
+    return this.next(request)
+  }
+}